Removing comments about alternate approach from code.
There was an alternate approach to make the test robust by not using the naked function.
A Equivalent 'C' function capable to achieve similar dissambly instructions and handle exception.
But it may use stack if no compiler optimisation is used, if stack used, then corruption could occur, therefore this solution was not chosen.

**--------**
__attribute__((noinline)) uint32_t call_mem(uint32_t addr)
{
    *(volatile uint32_t *)addr;
}
**--------**