Add PSA TF-M utility scripts in python system path
When python3 is enforced to build ARM_MUSCA_A1 target, it is unable to
find binary utility tools scripts which are imported from TF-M.

The reason for python3 environment is as follows:
Mbed OS + TFM contained a faulty boot record TLV, which failed the
attestation test (TF-M regression). The data in the boot record TLV
will be included in the generated attestation token as 1 item
in the SW_COMPONENTS claim.

This data (in the boot record TLV) is pre encoded in CBOR format at
build time and appended to the image during the image signing process
(done by the imgtool Python3 script).

Signed-off-by: Vikas Katariya <vikas.katariya@arm.com>